Refractored location of util functions and minor changes
