--- fairing/ml_tasks/__init__.py
@@ -1,1 +1,0 @@
-from .tasks import TrainJob, PredictionEndpoint
